requirements.txt
pandas>=1.3.0
plotly>=5.3.0
seaborn>=0.11.0
matplotlib>=3.4.0
IMDbPY>=2021.4.18
requests>=2.26.0 